pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["SvamiBog <rubanezmaxim@gmailcom>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
sqlalchemy = "^2.0.30"
gunicorn = "^22.0.0"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.3"
validators = "^0.20.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
